Move dependencies back to Dependencies.scala
