+++ project/Versions.scala
@@ -1,0 +1,14 @@
+object Versions {
+  val CatsVersion           = "1.6.0"
+  val CirceFs2Version       = "0.11.0"
+  val CirceVersion          = "0.11.1"
+  val GeoTrellisVersion     = "3.0.0-M3"
+  val RefinedVersion        = "0.9.3"
+  val ScapegoatVersion      = "1.3.8"
+  val ShapelessVersion      = "2.3.3"
+  val spdxCheckerVersion    = "1.0.0"
+  val scalacheckCatsVersion = "0.1.1"
+  val scalatestVersion      = "3.0.4"
+  val sprayVersion          = "1.3.4"
+  val scalacheckVersion     = "1.14.0"
+}
